fix: reverse guard against null driver to fix tests failing
By adding the guard statement in the first place we couldn't explicitly set the
driver to null, thus creating a temporal coupling between the test methods
for the DriverManager. Removing the guard against null from the Current
setter is OK for now. Explicitly setting the driver to null will cause a bunch
of NoDriverExceptions anyway, so that's stupid.